# [Ubuntu] Disabled .NET warmup 