# Description

We have investigated and found out that installation of .NET is hugest bottleneck in image-generation of  Ubuntu. The Ubuntu `dotnetcore-sdk.sh` installation script takes 1h20m (the total time of image-generation is 3h30m). 

There are two reasons for it:

1. We install a lot of .NET versions: 2.1, 3.1, 5.0 on Ubuntu
2. We run a huge validation after installation that takes the most of the time

On Ubuntu, we run the following code for every version:
```
    dotnet new globaljson --sdk-version "$sdk"
    dotnet new "$sample"
    dotnet restore
    dotnet build
```

In scope of the investigation ([Speed up .NET installation process](https://github.com/actions/virtual-environments/discussions/2367)) we decided to remove warmup stage  , since there are no performance benefits at all. 

Platform | Downloading and installation time with warmup | Downloading and installation time without warmup
-- | -- | --
Ubuntu | 1h40m | 15m


#### Related issue:
https://github.com/actions/virtual-environments-internal/issues/1988

## Check list
- [x] Related issue / work item is attached
- [ ] Tests are written (if applicable)
- [ ] Documentation is updated (if applicable)
- [ ] Changes are tested and related VM images are successfully generated